feat: Add doctor admitted patient list functionality to doctor opd&IPD portal
This commit adds the necessary code changes to implement the functionality for fetching and displaying the list of admitted patients and out door patients in the doctor IPD&OPD portal. It includes the addition of the `getDoctorAdmittedPatientList` &`getDoctorConsultationList` methods to the `DocPortalRepository` interface and the corresponding implementation in the `DocPortalRepositoryImpl` class. The `DoctorAdmittedPatientBloc` and related classes are also added to handle the state management for fetching the admitted& out door patient list.