feat: Add appRegistration method to AppRepositoryImpl
This commit adds the appRegistration method to the AppRepositoryImpl class. The method allows for registering a new user with the provided user details. If the registration is successful, the user is added to the database. If not, an error is thrown. This addition enhances the functionality of the app registration feature.