chore: Update Android namespace and signing configuration
This commit updates the Android namespace in the build.gradle file to "com.delta.hospital" for better package naming. It also adds the signing configuration for the release build in the build.gradle file, using the properties specified in the key.properties file. These changes improve the project's organization and prepare it for release.