feat: Update Android compileSdk and sourceCompatibility versions
This commit updates the Android `compileSdk` version to `35` and the `sourceCompatibility` version to `JavaVersion.VERSION_17` in the `build.gradle` file. The `compileSdk` version is updated to ensure compatibility with the latest Android platform, while the `sourceCompatibility` version is updated to utilize Java 17 features. These changes improve the project's compatibility and leverage the latest language features.