feat: Add findRegAndCreateUser method to PatPortalRemoteDataSource and PatPortalRepository
This commit adds the `findRegAndCreateUser` method to the `PatPortalRemoteDataSource` and `PatPortalRepository` classes. The method allows finding a patient's registration and creating a user based on their MRN. This enhancement improves the functionality of the patient portal by providing a convenient way to search for registration information and create a user account.