feat: Add isDense property to CommonTextFieldWidget
This commit adds the `isDense` property to the `CommonTextFieldWidget` in order to control the density of the input field. By setting `isDense` to `true`, the input field will have a compact appearance. This improvement enhances the user experience by allowing for more efficient use of screen space in forms or layouts where multiple input fields are present.