feat: Reset consultation type and patient user when changing patient type
This commit modifies the `_BookAppointmentViewState` class in the `book_appointment_view.dart` file. It adds a reset method for the `ConsultationType` and `PatientPortalUser` variables in the `VariableStateCubit` when the patient type is changed. This enhancement ensures that the consultation type and patient user are reset when switching between different patient types, improving the accuracy of the selected values during the appointment booking process.