feat: Add onTap property to CommonTextFieldWidget
This commit adds the `onTap` property to the `CommonTextFieldWidget` in order to handle tap events on the input field. By providing a callback function to the `onTap` property, developers can specify custom behavior when the input field is tapped. This enhancement improves the flexibility and interactivity of the `CommonTextFieldWidget` component.