feat: Add PaymodeListResponse model and API method
This commit adds the PaymodeListResponse model and the getPayModeList API method to the AppRemoteDataSource. The PaymodeListResponse model represents the response data for retrieving the list of paymodes from the server. The getPayModeList method allows fetching the list of available paymodes. This feature enhances the functionality of the app by providing access to the available paymodes for online appointments.