feat: Remove unused import statements and fix bug in cart view
The code changes in this commit involve removing unused import statements in the `items_booking_repository_impl.dart` file and fixing a bug in the `cart_view.dart` file. The bug fix includes removing the `dart:developer` import statement and ensuring the correct functionality of the `CartView` component. This commit improves code cleanliness and resolves a bug in the cart view.