feat: Add getPrescriptionPdf method to PatPortalRepository
This commit adds the `getPrescriptionPdf` method to the `PatPortalRepository` interface. The method takes in a prescription ID and a token as parameters and returns a `Uint8List` representing the PDF file of the prescription. This change enhances the functionality of the repository by allowing users to retrieve prescription PDFs from the remote data source.