feat: Add reset functionality to CommonDropdownButton
This commit adds a new `reset` parameter to the `CommonDropdownButton` widget, allowing users to reset the selected value by tapping on an icon. The reset functionality is implemented by calling the `reset` callback function when the icon is tapped. This enhancement improves the user experience by providing an easy way to clear the selected value in the dropdown button.